pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ip-whois-tool"
dynamic = ["version"]
description = "A tool for looking up WHOIS information for IP addresses"
readme = "README.md"
authors = [
    {name = "IP WHOIS Tool Contributors", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
keywords = ["whois", "ip", "networking", "rdap", "dns"]
dependencies = [
    "ipwhois==1.2.0",
    "python-whois>=0.9.5",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "typing-extensions>=4.4.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ip_whois_tool"
"Bug Tracker" = "https://github.com/yourusername/ip_whois_tool/issues"
"Source Code" = "https://github.com/yourusername/ip_whois_tool"
"Documentation" = "https://github.com/yourusername/ip_whois_tool#readme"

[project.scripts]
ip-whois = "whois_tool.cli:main"
ip-lookup = "whois_tool.cli:main"

[tool.setuptools]
packages = ["whois_tool"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "whois_tool.__version__"}

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I", "N", "B", "UP", "C", "A"]
ignore = ["E203", "E501"]
exclude = [
    ".git",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
    "venv",
    ".venv",
]

[tool.ruff.isort]
known-first-party = ["whois_tool"]
combine-as-imports = true

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["whois_tool"]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=whois_tool --cov-report=term-missing -v"

[tool.coverage.run]
source = ["whois_tool"]
omit = ["*/tests/*", "*/venv/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

